SPARK_UNN/MLCube/sparkUNN_mlcube/project/requirements.txt
# Include all your requirements here
typer
torchmetrics==0.11.4
git+https://github.com/NVIDIA/dllogger
git+https://github.com/NVIDIA/mlperf-common.git
nibabel==3.2.1
joblib==1.0.1
pytorch-lightning==1.7.7
scikit-learn==1.0
scikit-image==0.18.3
scipy==1.8.1
rich==12.5.0